requirements.txt
streamlit
fastapi
uvicorn
pypdf
langchain
sentence-transformers
faiss-cpu
transformers
torch
